+++ packages/hooks/index.tsx
@@ -1,5 +1,4 @@
 export { default as useInterval } from './useInterval';
 export { default as useMedia } from './useMedia';
 export { default as useMediaQuery } from './useMediaQuery';
-export { default as useMarketUrlQuery } from './useMarketUrlQuery';
 export { default as useUrlQuery } from './useUrlQuery';
